+++ src/sdc/writer/__init__.py
@@ -9,3 +9,4 @@ from ._json import JsonSampleDataWriter
 from ._multi import MultiWriter
 from ._nir import NIRWriter
 from ._pyfunc import PythonFunctionWriter
+from ._zip import ZipWriter