+++ src/sdc/writer/_adams.py
@@ -2,17 +2,17 @@ import argparse
 import os
 from typing import List, Dict, Any
 
-from javaproperties import Properties, dump
+from javaproperties import Properties, dump, dumps
 from seppl.placeholders import placeholder_list, InputBasedPlaceholderSupporter
 from seppl.io import DirectStreamWriter
 from wai.logging import LOGGING_WARNING
 from wai.spectralio.adams import DATATYPE_SUFFIX
 from wai.spectralio.adams import Writer as SWriter
 
-from sdc.api import Spectrum2D, SplittableStreamWriter, SplittableSampleDataStreamWriter, make_list, SpectralIOWriter
+from sdc.api import Spectrum2D, SplittableStreamWriter, SplittableSampleDataStreamWriter, make_list, SpectralIOWriter, DefaultExtensionWriter
 
 
-class AdamsWriter(SplittableStreamWriter, SpectralIOWriter, DirectStreamWriter, InputBasedPlaceholderSupporter):
+class AdamsWriter(SplittableStreamWriter, SpectralIOWriter, DirectStreamWriter, DefaultExtensionWriter, InputBasedPlaceholderSupporter):
 
     def __init__(self, output_dir: str = None, output_sampledata: bool = None,
                  split_names: List[str] = None, split_ratios: List[int] = None, split_group: str = None,
@@ -58,6 +58,16 @@ class AdamsWriter(SplittableStreamWriter, SpectralIOWriter, DirectStreamWriter, 
         """
         return "Saves the spectrum in ADAMS .spec format."
 
+    @property
+    def default_extension(self) -> str:
+        """
+        Returns the default extension (incl dot) for this file type.
+
+        :return: the default extension
+        :rtype: str
+        """
+        return ".spec"
+
     def _create_argparser(self) -> argparse.ArgumentParser:
         """
         Creates an argument parser. Derived classes need to fill in the options.
@@ -66,7 +76,7 @@ class AdamsWriter(SplittableStreamWriter, SpectralIOWriter, DirectStreamWriter, 
         :rtype: argparse.ArgumentParser
         """
         parser = super()._create_argparser()
-        parser.add_argument("-o", "--output", type=str, help="The directory to store the .spec files in. Any defined splits get added beneath there. " + placeholder_list(obj=self), required=True)
+        parser.add_argument("-o", "--output", type=str, help="The directory to store the .spec files in. Any defined splits get added beneath there. " + placeholder_list(obj=self), required=False)
         parser.add_argument("--output_sampledata", action="store_true", help="Outputs the sampledata as well")
         return parser
 
@@ -127,6 +137,9 @@ class AdamsWriter(SplittableStreamWriter, SpectralIOWriter, DirectStreamWriter, 
 
         :param data: the data to write (single record or iterable of records)
         """
+        if self.output_dir is None:
+            raise Exception("No output directory specified!")
+
         for item in make_list(data):
             sub_dir = self.session.expand_placeholders(self.output_dir)
             if self.splitter is not None:
@@ -141,17 +154,19 @@ class AdamsWriter(SplittableStreamWriter, SpectralIOWriter, DirectStreamWriter, 
             self.logger().info("Writing spectrum to: %s" % path)
             self._writer.write([item.spectrum], path)
 
-    def write_stream_fp(self, data, fp):
+    def write_stream_fp(self, data, fp, as_bytes: bool):
         """
         Saves the data one by one.
 
         :param data: the data to write (single record or iterable of records)
         :param fp: the file-like object to write to
+        :param as_bytes: whether to write as str or bytes
+        :type as_bytes: bool
         """
-        self._writer.write_fp([x.spectrum for x in make_list(data)], fp, False)
+        self._writer.write_fp([x.spectrum for x in make_list(data)], fp, as_bytes)
 
 
-class ReportSampleDataWriter(SplittableSampleDataStreamWriter, DirectStreamWriter, InputBasedPlaceholderSupporter):
+class ReportSampleDataWriter(SplittableSampleDataStreamWriter, DirectStreamWriter, DefaultExtensionWriter, InputBasedPlaceholderSupporter):
 
     def __init__(self, output_dir: str = None,
                  split_names: List[str] = None, split_ratios: List[int] = None, split_group: str = None,
@@ -193,6 +208,16 @@ class ReportSampleDataWriter(SplittableSampleDataStreamWriter, DirectStreamWrite
         """
         return "Saves the sample data in ADAMS .report format."
 
+    @property
+    def default_extension(self) -> str:
+        """
+        Returns the default extension (incl dot) for this file type.
+
+        :return: the default extension
+        :rtype: str
+        """
+        return ".report"
+
     def _create_argparser(self) -> argparse.ArgumentParser:
         """
         Creates an argument parser. Derived classes need to fill in the options.
@@ -201,7 +226,7 @@ class ReportSampleDataWriter(SplittableSampleDataStreamWriter, DirectStreamWrite
         :rtype: argparse.ArgumentParser
         """
         parser = super()._create_argparser()
-        parser.add_argument("-o", "--output", type=str, help="The directory to store the .report files in. Any defined splits get added beneath there. " + placeholder_list(obj=self), required=True)
+        parser.add_argument("-o", "--output", type=str, help="The directory to store the .report files in. Any defined splits get added beneath there. " + placeholder_list(obj=self), required=False)
         return parser
 
     def _apply_args(self, ns: argparse.Namespace):
@@ -246,6 +271,9 @@ class ReportSampleDataWriter(SplittableSampleDataStreamWriter, DirectStreamWrite
 
         :param data: the data to write (single record or iterable of records)
         """
+        if self.output_dir is None:
+            raise Exception("No output directory specified!")
+
         for item in make_list(data):
             sub_dir = self.session.expand_placeholders(self.output_dir)
             if self.splitter is not None:
@@ -261,14 +289,19 @@ class ReportSampleDataWriter(SplittableSampleDataStreamWriter, DirectStreamWrite
             with open(path, "w") as fp:
                 dump(self._to_props(item.sampledata), fp)
 
-    def write_stream_fp(self, data, fp):
+    def write_stream_fp(self, data, fp, as_bytes: bool):
         """
         Saves the data one by one.
 
         :param data: the data to write (single record or iterable of records)
         :param fp: the file-like object to write to
+        :param as_bytes: whether to write as str or bytes
+        :type as_bytes: bool
         """
         data = make_list(data)
         if len(data) != 1:
             raise Exception("Can only save single sample data at a time!")
-        dump(self._to_props(data[0].sampledata), fp)
+        if as_bytes:
+            fp.write(dumps(self._to_props(data[0].sampledata)).encode())
+        else:
+            dump(self._to_props(data[0].sampledata), fp)