+++ src/sdc/writer/_csv.py
@@ -1,5 +1,6 @@
 import argparse
 import csv
+from io import StringIO
 from typing import List
 
 from seppl.placeholders import InputBasedPlaceholderSupporter
@@ -7,10 +8,11 @@ from seppl.io import DirectBatchWriter
 from wai.logging import LOGGING_WARNING
 from wai.spectralio.csv import Writer as SWriter, PLACEHOLDERS, PH_WAVE_NUMBER
 
-from sdc.api import Spectrum2D, SplittableBatchWriter, SplittableSampleDataBatchWriter, SampleData, SAMPLE_ID, SpectralIOWriter
+from sdc.api import Spectrum2D, SplittableBatchWriter, SplittableSampleDataBatchWriter, SampleData, SAMPLE_ID, \
+    SpectralIOWriter, make_list, DefaultExtensionWriter
 
 
-class CSVWriter(SplittableBatchWriter, SpectralIOWriter, DirectBatchWriter, InputBasedPlaceholderSupporter):
+class CSVWriter(SplittableBatchWriter, SpectralIOWriter, DirectBatchWriter, DefaultExtensionWriter, InputBasedPlaceholderSupporter):
 
     def __init__(self, output_file: str = None, sample_id: str = None, sample_data: List[str] = None,
                  sample_data_prefix: str = None, wave_numbers_format: str = None,
@@ -66,6 +68,16 @@ class CSVWriter(SplittableBatchWriter, SpectralIOWriter, DirectBatchWriter, Inpu
         """
         return "Saves the spectra in CSV format (row-wise)."
 
+    @property
+    def default_extension(self) -> str:
+        """
+        Returns the default extension (incl dot) for this file type.
+
+        :return: the default extension
+        :rtype: str
+        """
+        return ".csv"
+
     def _create_argparser(self) -> argparse.ArgumentParser:
         """
         Creates an argument parser. Derived classes need to fill in the options.
@@ -74,7 +86,7 @@ class CSVWriter(SplittableBatchWriter, SpectralIOWriter, DirectBatchWriter, Inpu
         :rtype: argparse.ArgumentParser
         """
         parser = super()._create_argparser()
-        parser.add_argument("-o", "--output", type=str, help="The CSV file to store the spectra in.", required=True)
+        parser.add_argument("-o", "--output", type=str, help="The CSV file to store the spectra in.", required=False)
         parser.add_argument("--sample_id", type=str, help="The name to use for the sample ID column.", required=False, default="sample_id")
         parser.add_argument("--sample_data", type=str, help="The sample data names to store in CSV file.", required=False, default=[], nargs="*")
         parser.add_argument("--sample_data_prefix", type=str, help="The prefix to use for the sample data columns.", required=False, default="")
@@ -149,22 +161,28 @@ class CSVWriter(SplittableBatchWriter, SpectralIOWriter, DirectBatchWriter, Inpu
         :param data: the data to write
         :type data: Iterable
         """
+        if self.output_file is None:
+            raise Exception("No output file specified!")
+
         output_file = self.session.expand_placeholders(self.output_file)
         self.logger().info("Writing spectra to: %s" % output_file)
         self._writer.write([x.spectrum for x in data], output_file)
 
-    def write_batch_fp(self, data, fp):
+    def write_batch_fp(self, data, fp, as_bytes: bool):
         """
         Saves the data in one go.
 
         :param data: the data to write
         :type data: Iterable
         :param fp: the file-like object to write to
+        :param as_bytes: whether to write as str or bytes
+        :type as_bytes: bool
         """
-        self._writer.write_fp([x.spectrum for x in data], fp, False)
+        data = make_list(data)
+        self._writer.write_fp([x.spectrum for x in data], fp, as_bytes)
 
 
-class CSVSampleDataWriter(SplittableSampleDataBatchWriter, DirectBatchWriter, InputBasedPlaceholderSupporter):
+class CSVSampleDataWriter(SplittableSampleDataBatchWriter, DirectBatchWriter, DefaultExtensionWriter, InputBasedPlaceholderSupporter):
 
     def __init__(self, output_file: str = None, sample_id: str = None, sample_data: List[str] = None,
                  sample_data_prefix: str = None,
@@ -216,6 +234,16 @@ class CSVSampleDataWriter(SplittableSampleDataBatchWriter, DirectBatchWriter, In
         """
         return "Saves the sample data in CSV format (row-wise)."
 
+    @property
+    def default_extension(self) -> str:
+        """
+        Returns the default extension (incl dot) for this file type.
+
+        :return: the default extension
+        :rtype: str
+        """
+        return ".csv"
+
     def _create_argparser(self) -> argparse.ArgumentParser:
         """
         Creates an argument parser. Derived classes need to fill in the options.
@@ -224,7 +252,7 @@ class CSVSampleDataWriter(SplittableSampleDataBatchWriter, DirectBatchWriter, In
         :rtype: argparse.ArgumentParser
         """
         parser = super()._create_argparser()
-        parser.add_argument("-o", "--output", type=str, help="The CSV file to store the sample data in.", required=True)
+        parser.add_argument("-o", "--output", type=str, help="The CSV file to store the sample data in.", required=False)
         parser.add_argument("--sample_id", type=str, help="The name to use for the sample ID column.", required=False, default="sample_id")
         parser.add_argument("--sample_data", type=str, help="The sample data names to store in CSV file.", required=False, default=[], nargs="*")
         parser.add_argument("--sample_data_prefix", type=str, help="The prefix to use for the sample data columns.", required=False, default="")
@@ -271,21 +299,32 @@ class CSVSampleDataWriter(SplittableSampleDataBatchWriter, DirectBatchWriter, In
         :param data: the data to write
         :type data: Iterable
         """
+        if self.output_file is None:
+            raise Exception("No output file specified!")
+
         output_file = self.session.expand_placeholders(self.output_file)
         self.logger().info("Writing sample data to: %s" % output_file)
 
         with open(output_file, "w") as fp:
-            self.write_batch_fp(data, fp)
+            self.write_batch_fp(data, fp, False)
 
-    def write_batch_fp(self, data, fp):
+    def write_batch_fp(self, data, fp, as_bytes: bool):
         """
         Saves the data in one go.
 
         :param data: the data to write
         :type data: Iterable
         :param fp: the file-like object to write to
+        :param as_bytes: whether to write as str or bytes
+        :type as_bytes: bool
         """
-        writer = csv.writer(fp, quoting=csv.QUOTE_MINIMAL)
+        data = make_list(data)
+        if as_bytes:
+            buffer = StringIO()
+            writer = csv.writer(buffer, quoting=csv.QUOTE_MINIMAL)
+        else:
+            buffer = None
+            writer = csv.writer(fp, quoting=csv.QUOTE_MINIMAL)
 
         # header
         row = [self.sample_id]
@@ -312,3 +351,6 @@ class CSVSampleDataWriter(SplittableSampleDataBatchWriter, DirectBatchWriter, In
                 else:
                     row.append("")
             writer.writerow(row)
+
+        if as_bytes:
+            fp.write(buffer.getvalue().encode())