+++ src/sdc/writer/_json.py
@@ -7,10 +7,10 @@ from seppl.placeholders import placeholder_list
 from seppl.io import DirectStreamWriter
 from wai.logging import LOGGING_WARNING
 
-from sdc.api import SplittableSampleDataStreamWriter, make_list
+from sdc.api import SplittableSampleDataStreamWriter, make_list, DefaultExtensionWriter
 
 
-class JsonSampleDataWriter(SplittableSampleDataStreamWriter, DirectStreamWriter):
+class JsonSampleDataWriter(SplittableSampleDataStreamWriter, DirectStreamWriter, DefaultExtensionWriter):
 
     def __init__(self, output_dir: str = None, indent: int = None,
                  split_names: List[str] = None, split_ratios: List[int] = None, split_group: str = None,
@@ -55,6 +55,16 @@ class JsonSampleDataWriter(SplittableSampleDataStreamWriter, DirectStreamWriter)
         """
         return "Saves the sample data in JSON format."
 
+    @property
+    def default_extension(self) -> str:
+        """
+        Returns the default extension (incl dot) for this file type.
+
+        :return: the default extension
+        :rtype: str
+        """
+        return ".json"
+
     def _create_argparser(self) -> argparse.ArgumentParser:
         """
         Creates an argument parser. Derived classes need to fill in the options.
@@ -63,7 +73,7 @@ class JsonSampleDataWriter(SplittableSampleDataStreamWriter, DirectStreamWriter)
         :rtype: argparse.ArgumentParser
         """
         parser = super()._create_argparser()
-        parser.add_argument("-o", "--output", type=str, help="The directory to store the .json files in. Any defined splits get added beneath there. " + placeholder_list(obj=self), required=True)
+        parser.add_argument("-o", "--output", type=str, help="The directory to store the .json files in. Any defined splits get added beneath there. " + placeholder_list(obj=self), required=False)
         parser.add_argument("--indent", type=int, help="The indent to use for pretty-printing the JSON instead of optimal file size.", default=None, required=False)
         return parser
 
@@ -84,6 +94,9 @@ class JsonSampleDataWriter(SplittableSampleDataStreamWriter, DirectStreamWriter)
 
         :param data: the data to write (single record or iterable of records)
         """
+        if self.output_dir is None:
+            raise Exception("No output directory specified!")
+
         for item in make_list(data):
             sub_dir = self.session.expand_placeholders(self.output_dir)
             if self.splitter is not None:
@@ -99,14 +112,19 @@ class JsonSampleDataWriter(SplittableSampleDataStreamWriter, DirectStreamWriter)
             with open(path, "w") as fp:
                 json.dump(item.sampledata, fp, indent=self.indent)
 
-    def write_stream_fp(self, data, fp):
+    def write_stream_fp(self, data, fp, as_bytes: bool):
         """
         Saves the data one by one.
 
         :param data: the data to write (single record or iterable of records)
         :param fp: the file-like object to write to
+        :param as_bytes: whether to write as str or bytes
+        :type as_bytes: bool
         """
         data = make_list(data)
         if len(data) != 1:
             raise Exception("Can only save single sample data at a time!")
-        json.dump(data[0].sampledata, fp, indent=self.indent)
+        if as_bytes:
+            fp.write(json.dumps(data[0].sampledata, indent=self.indent).encode())
+        else:
+            json.dump(data[0].sampledata, fp, indent=self.indent)