+++ src/sdc/writer/_nir.py
@@ -6,10 +6,10 @@ from seppl.io import DirectBatchWriter
 from wai.logging import LOGGING_WARNING
 from wai.spectralio.nir import Writer as SWriter
 
-from sdc.api import Spectrum2D, SplittableBatchWriter, SpectralIOWriter
+from sdc.api import Spectrum2D, SplittableBatchWriter, SpectralIOWriter, DefaultExtensionWriter, make_list
 
 
-class NIRWriter(SplittableBatchWriter, SpectralIOWriter, DirectBatchWriter, InputBasedPlaceholderSupporter):
+class NIRWriter(SplittableBatchWriter, SpectralIOWriter, DirectBatchWriter, DefaultExtensionWriter, InputBasedPlaceholderSupporter):
 
     def __init__(self, output_file: str = None, instrument_name: str = None,
                  product_code: str = None, product_code_from_field: bool = None, client: str = None,
@@ -108,6 +108,16 @@ class NIRWriter(SplittableBatchWriter, SpectralIOWriter, DirectBatchWriter, Inpu
         """
         return "Saves the spectra in FOSS NIR format."
 
+    @property
+    def default_extension(self) -> str:
+        """
+        Returns the default extension (incl dot) for this file type.
+
+        :return: the default extension
+        :rtype: str
+        """
+        return ".nir"
+
     def _create_argparser(self) -> argparse.ArgumentParser:
         """
         Creates an argument parser. Derived classes need to fill in the options.
@@ -116,7 +126,7 @@ class NIRWriter(SplittableBatchWriter, SpectralIOWriter, DirectBatchWriter, Inpu
         :rtype: argparse.ArgumentParser
         """
         parser = super()._create_argparser()
-        parser.add_argument("-o", "--output", type=str, help="The NIR file to store the spectra in.", required=True)
+        parser.add_argument("-o", "--output", type=str, help="The NIR file to store the spectra in.", required=False)
         parser.add_argument("--instrument_name", type=str, help="The instrument name to use in the header", required=False, default="<not implemented>")
         parser.add_argument("--product_code", type=str, help="Either the attribute name with the product code in it, or the actual product code to be used", required=False, default="01")
         parser.add_argument("--product_code_from_field", action="store_true", help="Whether to use the product code option as the attribute name containing the actual product code")
@@ -264,16 +274,22 @@ class NIRWriter(SplittableBatchWriter, SpectralIOWriter, DirectBatchWriter, Inpu
         :param data: the data to write
         :type data: Iterable
         """
+        if self.output_file is None:
+            raise Exception("No output file specified!")
+
         output_file = self.session.expand_placeholders(self.output_file)
         self.logger().info("Writing spectra to: %s" % output_file)
         self._writer.write([x.spectrum for x in data], output_file)
 
-    def write_batch_fp(self, data, fp):
+    def write_batch_fp(self, data, fp, as_bytes: bool):
         """
         Saves the data in one go.
 
         :param data: the data to write
         :type data: Iterable
         :param fp: the file-like object to write to
+        :param as_bytes: whether to write as str or bytes
+        :type as_bytes: bool
         """
-        self._writer.write_fp([x.spectrum for x in data], fp, True)
+        data = make_list(data)
+        self._writer.write_fp([x.spectrum for x in data], fp, as_bytes)